+++ packages/patrol/example/integration_test/config.dart
@@ -1,7 +1,10 @@
 import 'package:patrol/patrol.dart';
 
-const patrolConfig = PatrolTesterConfig(
-  appName: 'Example App',
+const patrolConfig = PatrolTesterConfig(appName: 'Example App');
+
+final nativeAutomatorConfig = NativeAutomatorConfig(
   packageName: 'pl.leancode.patrol.example',
   bundleId: 'pl.leancode.patrol.Example',
 );
+
+const hostAutomatorConfig = HostAutomatorConfig();