+++ packages/patrol/lib/src/custom_finders/custom_finders.dart
@@ -1,5 +1,4 @@
 export 'common.dart';
 export 'exceptions.dart';
 export 'patrol_finder.dart';
-export 'patrol_test_config.dart';
 export 'patrol_tester.dart';