+++ packages/patrol/lib/src/custom_finders/patrol_tester.dart
@@ -1,7 +1,63 @@
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:patrol/patrol.dart';
-import 'package:patrol/src/host/host_automator.dart';
+
+/// Common configuration for [PatrolTester] and [PatrolFinder].
+class PatrolTesterConfig {
+  /// Creates a new [PatrolTesterConfig].
+  const PatrolTesterConfig({
+    this.existsTimeout = const Duration(seconds: 10),
+    this.visibleTimeout = const Duration(seconds: 10),
+    this.settleTimeout = const Duration(seconds: 10),
+    this.andSettle = true,
+    this.appName,
+  });
+
+  /// Time after which [PatrolFinder.waitUntilExists] fails if it doesn't finds
+  /// a widget.
+  final Duration existsTimeout;
+
+  /// Time after which [PatrolFinder.waitUntilVisible] fails if it doesn't finds
+  /// a widget.
+  ///
+  /// [PatrolFinder.waitUntilVisible] is used internally by [PatrolFinder.tap]
+  /// and [PatrolFinder.enterText].
+  ///
+  final Duration visibleTimeout;
+
+  /// Time after which [PatrolTester.pumpAndSettle] fails.
+  final Duration settleTimeout;
+
+  /// Whether to call [WidgetTester.pumpAndSettle] after actions such as
+  /// [PatrolFinder.tap] and [PatrolFinder]. If false, only [WidgetTester.pump]
+  /// is called.
+  final bool andSettle;
+
+  /// Name of the application under test.
+  ///
+  /// Used in [PatrolTester.log].
+  final String? appName;
+
+  /// Creates a copy of this config but with the given fields replaced with the
+  /// new values.
+  PatrolTesterConfig copyWith({
+    Duration? existsTimeout,
+    Duration? visibleTimeout,
+    Duration? settleTimeout,
+    bool? andSettle,
+    String? appName,
+    String? packageName,
+    String? bundleId,
+  }) {
+    return PatrolTesterConfig(
+      existsTimeout: existsTimeout ?? this.existsTimeout,
+      visibleTimeout: visibleTimeout ?? this.visibleTimeout,
+      settleTimeout: settleTimeout ?? this.settleTimeout,
+      andSettle: andSettle ?? this.andSettle,
+      appName: appName ?? this.appName,
+    );
+  }
+}
 
 /// Default amount to drag by when scrolling.
 const defaultScrollDelta = 64.0;
@@ -60,7 +116,7 @@ class PatrolTester {
   });
 
   /// Global configuration of this tester.
-  final PatrolTestConfig config;
+  final PatrolTesterConfig config;
 
   /// Flutter's widget tester that this [PatrolTester] wraps.
   final WidgetTester tester;